Change data type of m_delayTime to qint64 to fix pc-lint warning

+++ HimalayaMain/Master/Components/Scheduler/Source/SchedulerMainThreadController.cpp
@@ -914,9 +914,9 @@ void SchedulerMainThreadController::CheckResuemFromPause(SchedulerStateMachine_t
     {        
         if (m_CurProgramStepIndex == 0)
         {
-            if ((now-m_PauseStartTime) > static_cast<qint64>(m_delayTime*1000))
+            if ((now-m_PauseStartTime) > m_delayTime*1000)
             {
-                offset = now - m_PauseStartTime - static_cast<qint64>(m_delayTime*1000);
+                offset = now - m_PauseStartTime - m_delayTime*1000;
                 m_CurProgramStepInfo.durationInSeconds -= m_delayTime;
                 m_delayTime = 0;
             }
@@ -939,9 +939,9 @@ void SchedulerMainThreadController::CheckResuemFromPause(SchedulerStateMachine_t
         {
             if (m_CurProgramStepIndex == 0)
             {
-                if ((now-m_PauseStartTime) > static_cast<qint64>(m_delayTime*1000))
+                if ((now-m_PauseStartTime) > m_delayTime*1000)
                 {
-                    offset = now - m_PauseStartTime - static_cast<qint64>(m_delayTime*1000);
+                    offset = now - m_PauseStartTime - m_delayTime*1000;
                     m_CurProgramStepInfo.durationInSeconds -= m_delayTime;
                     m_delayTime = 0;
                 }
